backend/requirements.txt
Django==2.1.2
django-cors-headers==2.4.0
django-extensions==2.1.3
djangorestframework==3.9.0
mysqlclient==1.3.12
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
six==1.11.0
